\name{c2i}
\alias{c2i}
\title{c2i}
\description{coordinates to index}
\usage{
c2i(nrows,
x,
y)
}
\arguments{
  \item{nrows}{number of rows in the matrix}
  \item{x}{the row coordinate}
  \item{y}{the col coordinate}
}

\value{the index into the matrix}

\author{David L Gibbs}





