\name{compareNetworksWithFishersExactTestProcona}
\alias{compareNetworksWithFishersExactTestProcona}
\title{compareNetworksWithFishersExactTestProcona}
\description{Convienence function for calling the compareNetworksWithFishersExactTest
using only two procona objects.  Plot xlab and ylab is taken from procona obj.}
\usage{
compareNetworksWithFishersExactTestProcona(netobj1,
netobj2,
title)
}
\arguments{
  \item{netobj1}{procona object for network 1}
  \item{netobj2}{procona object for network 2}
  \item{title}{plot title}
}

\value{returns the list of fisher -log pvalues, and overlaps between modules}

\author{David L Gibbs}





