\name{getPeptideNAs}
\alias{getPeptideNAs}
\title{getPeptideNAs}
\description{This function returns the number of NAs for each peptide.}
\usage{
getPeptideNAs(pepdat)
}
\arguments{
  \item{pepdat}{the peptide data.}
}

\value{returns a list of counts of NAs for each peptide.}

\author{David L Gibbs}





