\name{hclust-class}
\Rdversion{1.1}
\docType{class}
\alias{hclust-class}

\title{Class \code{"hclust"}}
\description{
From the OneHandClapping package. Thanks! Dummy class to permit object of S3 class \code{hclust} in S4 class definition of Screening
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("hclust", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{list}"}, from data part.
Class \code{"\linkS4class{vector}"}, by class "list", distance 2.
}
\section{Methods}{
No methods defined with class "hclust" in the signature.
}
\note{
This class is just defined as a dummy class. No objects should be instantiated.
}

\section{Warning }{This class is just defined as a dummy class. No objects should be instantiated.}

\examples{
showClass("hclust")
}
\keyword{classes}
