\name{masstagdb}
\alias{masstagdb}
\title{A simulated mass tag data base}
\description{
  The mass tag database, which would be used to identify peptides, simply
  maps peptide IDs to peptide sequences and protein matches.
}
\usage{data(ProCoNA_Data)}
\format{A data frame with peptide IDs, peptide sequences, and protein matches.}
\keyword{datasets}
