\name{mergedMEs}
\alias{mergedMEs}
\title{mergedMEs}
\description{The module eigenvectors before the module merging step.}
\usage{mergedMEs(p)}
\arguments{
  \item{p}{The procona network object.}
}

\value{Returns the matrix of module eigenvectors, before module merging, in columns.}

\author{David L Gibbs}





