\name{MMvsPSallModules}
\alias{MMvsPSallModules}
\title{MMvsPSallModules}
\description{Produce pdfs for all modules.}
\usage{
MMvsPSallModules(net,
peptable,
phenoVec,
prefixName = "mm_vs_ps_")
}
\arguments{
  \item{net}{the procona network object}
  \item{peptable}{the peptide data}
  \item{phenoVec}{the phenotypic trait, as a numeric vector}
  \item{prefixName}{the}
}

\value{nothing returned}

\author{David L Gibbs}





