\name{name}
\alias{name}
\title{name}
\description{The name of the proconaNet.}
\usage{name(p)}
\arguments{
  \item{p}{the proconaNet obj}
}

\value{Returns the name of the network.}


\examples{
data(ProCoNA_Data)
net1 <- buildProconaNetwork("peptide network", peptideData)
name(net1)
}
\author{David L Gibbs}





