\name{orderMatrixIndex}
\alias{orderMatrixIndex}
\title{orderMatrixIndex}
\description{Order the the matrix by upper diag in a greedy fashion}
\usage{orderMatrixIndex(mat)}
\arguments{
  \item{mat}{A matrix}
}

\value{returns a matrix in order of greatest in upper diagonal direction.}

\author{David L Gibbs}





