\name{pepnet-class}
\Rdversion{1.1}
\docType{class}
\alias{pepnet-class}
\alias{pepnet}

\title{pepnet S4 class}
\description{NA}
\section{Objects from the Class}{Objects can be created by calls of the form \code{new(pepnet ...)}}
\section{Slots}{
  \describe{
    \item{\code{proconaVersion}:}{Object of class \code{"character"}}
    \item{\code{networkName}:}{Object of class \code{"character"}}
    \item{\code{samples}:}{Object of class \code{"character"}}
    \item{\code{adjacency}:}{Object of class \code{"matrix"}}
    \item{\code{TOM}:}{Object of class \code{"matrix"}}
    \item{\code{peptides}:}{Object of class \code{"character"}}
    \item{\code{pepTree}:}{Object of class \code{"hclust"}}
    \item{\code{dynamicColors}:}{Object of class \code{"numeric"}}
    \item{\code{MEs}:}{Object of class \code{"data.frame"}}
    \item{\code{mergedMEs}:}{Object of class \code{"data.frame"}}
    \item{\code{mergedColors}:}{Object of class \code{"numeric"}}
    \item{\code{colorOrder}:}{Object of class \code{"character"}}
    \item{\code{power}:}{Object of class \code{"numeric"}}
    \item{\code{networkType}:}{Object of class \code{"character"}}
    \item{\code{permtest}:}{Object of class \code{"matrix"}}
  }
}
\section{Methods}{
No methods defined with class "pepnet" in the signature.
}

\author{David L Gibbs}





