\name{printNet}
\alias{printNet}
\title{printNet}
\description{Prints general information about the network object.}
\usage{printNet(object)}
\arguments{
  \item{object}{The procona network object.}
}

\value{None.}

\examples{
\dontrun{
data(ProCoNA_Data)
net1 <- buildProconaNetwork("peptide network", peptideData)
printNet(net1)
}
}

\author{David L Gibbs}





