\name{procona-package}
\alias{procona-package}
\docType{package}
\title{Peptide co-expression network construction.}
\description{Peptide co-expression network construction, analysis, and visualization.}
\details{
\tabular{ll}{Package: \tab procona\cr
Type: \tab Package\cr
Title: \tab Peptide co-expression network construction.\cr
Version: \tab 0.13\cr
Date: \tab 2011-08-10\cr
Author: \tab David L Gibbs\cr
Maintainer: \tab David Gibbs <gibbsd@ohsu.edu>\cr
License: \tab GPLv3\cr
LazyLoad: \tab yes\cr
Depends: \tab WGCNA, GOstats, multicore\cr}
}
\author{David L Gibbs}

\keyword{ package }


