\name{proconaMergedColors}
\alias{proconaMergedColors}
\title{proconaMergedColors}
\description{this function returns the vector of module identifiers}
\usage{proconaMergedColors(p)}
\arguments{
  \item{p}{the procona obj}
}

\value{returns the vector module names for each peptide}

\author{David L Gibbs}





