\name{proconaMergedMEs}
\alias{proconaMergedMEs}
\title{proconaMergedMEs}
\description{this function returns the merged module eigenvectors}
\usage{proconaMergedMEs(p)}
\arguments{
  \item{p}{the procona obj}
}

\value{returns the matrix of module eigenvectors in columns}

\author{David L Gibbs}





