\name{proconaPeptides}
\alias{proconaPeptides}
\title{proconaPeptides}
\description{this function returns the names of the peptides in the pepnet object}
\usage{proconaPeptides(p)}
\arguments{
  \item{p}{the procona obj}
}

\value{returns the vector of peptide IDs}

\author{David L Gibbs}





