\name{proconaPermtest}
\alias{proconaPermtest}
\title{proconaPermtest}
\description{this function returns the names of the peptides in the pepnet object}
\usage{proconaPermtest(p)}
\arguments{
  \item{p}{the procona obj}
}

\value{returns the vector of peptide IDs}

\author{David L Gibbs}





