\name{subsetModCors}
\alias{subsetModCors}
\title{subsetModCors}
\description{subsets the module-phenotype correlation matrix which has funny rownames}
\usage{subsetModCors(modCors, modules)}
\arguments{
  \item{modCors}{The matrix of module-phenotype correlations}
  \item{modules}{Which modules are desired.}
}



\author{David L Gibbs}





