\name{subsetPeptideData}
\alias{subsetPeptideData}
\title{subsetPeptideData}
\description{Return the smaller peptide table, omitting NAs}
\usage{
subsetPeptideData(pepdat,
numNAsAllowed = NULL,
percentageNAsAllowed = 0.05)
}
\arguments{
  \item{pepdat}{the peptide information}
  \item{numNAsAllowed}{the number of NAs allowed for each peptide}
  \item{percentageNAsAllowed}{the percentage of NAs allowed for each peptide.}
}

\value{Returns the subset table.}

\author{David L Gibbs}





