\name{tukeyDataQuality}
\alias{tukeyDataQuality}
\title{tukeyDataQuality}
\description{Check data quality by comparing Tukey and Pearson.}
\usage{
tukeyDataQuality(pepdat)
}
\arguments{
  \item{pepdat}{peptide data}
}

\value{returns names of peptides that are questionable.}

\author{David L Gibbs}





