\name{universe}
\alias{universe}
\title{Entrez IDs for each Reference in the mass tag database.}
\description{
  The mass tag database, which would be used to identify peptides, simply
  maps peptide IDs to peptide sequences and protein matches. This represents 
  a mapping to Entrez IDs.
}
\usage{data(ProCoNA_Data)}
\format{A data frame with columns from (Uniprot IDs) and to (Entrez IDs).}
\keyword{datasets}
