\name{correlationWithPhenotypesHeatMap}
\alias{correlationWithPhenotypesHeatMap}
\title{correlationWithPhenotypesHeatMap}
\description{The heatmap showing the relation of the
modules with the phenotypes computed by
using the function relateModules(...)}
\usage{
correlationWithPhenotypesHeatMap(pnet,
pdat,
phenotypes,
plotName = "phenotypeHeatmap.pdf",
title = "Module-trait relationships")
}
\arguments{
  \item{pnet}{the peptide network object}
  \item{pdat}{the peptide data}
  \item{phenotypes}{matrix of phenotypic traits}
  \item{plotName}{the name of the saved plot}
  \item{title}{plot main title}
}

\value{the module eigenvector correlations}

\author{David L Gibbs}





