\name{correlationWithPhenotypesHeatMap}
\alias{correlationWithPhenotypesHeatMap}
\title{correlationWithPhenotypesHeatMap}
\description{The heatmap showing the relation of the
modules with the phenotypes computed by
using the function relateModules(...)}
\usage{correlationWithPhenotypesHeatMap(pnet, pdat, phenotypes, modules, 
    plotName = NULL, title = "Module-trait relationships", textSize = 0.4)}
\arguments{
  \item{pnet}{the peptide network object}
  \item{pdat}{the peptide data}
  \item{phenotypes}{matrix of phenotypic traits}
  \item{modules}{the vector of modules to plot}
  \item{plotName}{the name of the saved plot, NULL to show on screen}
  \item{title}{plot main title}
  \item{textSize}{
}
}

\value{the module eigenvector correlations}

\author{David L Gibbs}





