\name{getFisherMatrix}
\alias{getFisherMatrix}
\title{getFisherMatrix}
\description{Returns the matrix of fisher pvalues}
\usage{getFisherMatrix(net1, net2, colors1, colors2)}
\arguments{
  \item{net1}{Net1 list of entities in the network (nodes of network 1)}
  \item{net2}{Net2 list of entities in the network (nodes of network 2)}
  \item{colors1}{the module assignments for network 1}
  \item{colors2}{the module assignments for network 2}
}

\value{returns the fisher test pvalues and overlaps }

\author{David L Gibbs}





