\name{mergedColors}
\alias{mergedColors}
\title{mergedColors}
\description{this function returns the vector of module identifiers}
\usage{
mergedColors(p)
}
\arguments{
  \item{p}{the procona obj}
}

\value{returns the vector module names for each peptide}

\author{David L Gibbs}





