\name{MMvsPS}
\alias{MMvsPS}
\title{MMvsPS}
\description{Plots the module membership against the peptide significance
for a given trait and module}
\usage{
MMvsPS(net,
peptable,
phenoVec,
mod)
}
\arguments{
  \item{net}{The procona network}
  \item{peptable}{the peptide data, with rows as samples and columns as peptides}
  \item{phenoVec}{the phenotypic trait, vector}
  \item{mod}{the module of interest}
}

\value{returns a list of module memberships and peptide significances.}

\author{David L Gibbs}





