\name{moduleMemberCorrelations}
\alias{moduleMemberCorrelations}
\title{moduleMemberCorrelations}
\description{Computes the relation between the modules and
the phenotypes.}
\usage{moduleMemberCorrelations(pnet, pepdat, phenotypes)}
\arguments{
  \item{pnet}{The peptide net object}
  \item{pepdat}{The peptide data}
  \item{phenotypes}{The matrix of traits}
}

\value{returns a data frame of peptide correlation stats.}

\author{David L Gibbs}





