\name{modulePhenotypeCorrelations}
\alias{modulePhenotypeCorrelations}
\title{modulePhenotypeCorrelations}
\description{Computes the relation between the modules and
the phenotypes.}
\usage{
modulePhenotypeCorrelations(pnet,
pepdat,
phenotypes)
}
\arguments{
  \item{pnet}{The peptide net object}
  \item{pepdat}{The peptide data}
  \item{phenotypes}{The matrix of traits}
}

\value{returns a list of module member stats.}

\author{David L Gibbs}





